Updated request and response files to follow Python naming convention and moved into api folder
